write_data/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
google-cloud-bigquery==3.25.0
google-auth==2.33.0
